FIx editor browsable
